Revert "Try defining WEBKIT_DLL everywhere again"
We can't do this.

When WEBKIT_DLL is defined, WEBKIT_EXPORT is defined to be
__declspec(dllexport) when WEBKIT_IMPLEMENTATION is defined and
__declspec(dllimport) when WEBKIT_IMPLEMENTATION is undefined. That's
pretty much the same as every other module. If we were treating this
like every other module, we'd define both WEBKIT_DLL and
WEBKIT_IMPLEMENTATION everywhere.

But what's different is that when WEBKIT_IMPLEMENTATION is defined, a
bunch of private-to-WebCore code gets included in WebKit's headers. This
breaks the build when other modules include these headers, because they
don't have access to WebCore's headers.

So instead we have to only define WEBKIT_DLL for projects where
WEBKIT_IMPLEMENTATION is also defined. (If we define it for any other
projects, we'll get linker warnings about importing local symbols,
because the WebKit symbols will be decorated with
__declspec(dllimport).) This means that when other modules use WebKit's
headers, they won't see any decoration on WebKit's symbols at all. That
is just fine. We really only need WebKit's own modules and modules
*outside of* libchromiumcontent to see the decorations (the former so
that the symbols will get exported at all, the latter so that calls into
chromiumcontent.dll will be efficient).

This reverts commit bd4f5f21b24faa4ec8b6d69d30975d25d6f4713c.