Fix Release linker error on Linux
We were getting a linker error about duplicate definitions of malloc(),
one in base/process_util_linux.cc and one in
base/allocator/debugallocation_shim.cc. The problem was that the Release
build was confused about whether to use TCMalloc or not due to the
win_use_allocator_shim gyp variable being set. It seems like a bug that
Chromium's gyp files check that variable on platforms other than
Windows, but we can work around it for now by only defining that
variable on Windows.