Use relative paths in filenames.gypi
gyp will convert relative paths to absolute paths for us if we store
them in appropriately-named variables. Using gyp's automatic path
conversion seems to work better when generating Xcode projects. (For
example, gyp was complaining about files from libchromiumcontent being
included twice in the project, but that error goes away with this
change.)

See
https://chromium.googlesource.com/external/gyp/+/master/docs/InputFormatReference.md#Pathname-Relativization
for the full explanation of gyp's behavior. The essence is that by
naming these variables with *_files we get the automatic path conversion
behavior.