Replace base::OS::FOpen with ::fopen in mksnapshot
Since libv8 is linked as a dynamic library with mksnapshot, and they do
not share a common standard library, passing FILE* would result in
invalid_paramter error on Windows.

So instead of calling base::OS::FOpen which returns a FILE* from libv8,
this patch calls ::fopen directly.

I don't know why this was not a problem before Chrome 56.