Patch V8: Make TryHandleSignal available on all POSIX platforms
It is needed to fix Chromium compilation.

The same change in V8 (almost the same) was done here:
https://chromium-review.googlesource.com/c/v8/v8/+/508838
to fix compilation of code added to Chromium here:
https://chromium-review.googlesource.com/c/chromium/src/+/541956

Node.js doesn't have this change and doesn't need it
because it has "V8_TARGET_ARCH_X64" defined for all build targets
but Chromium has it only for V8 and not content/renderer
where "include/v8.h" header is used.