[webrtc][mac] cherry-pick fix for laggy mouse during desktop capture
The mouse is being capture twice, one time with a custom way
and one time with a built-in Mac API. The later was not known
so the mouse was wrongly captured twice.
Also the later way causes the cursor to move slowly on the
machine that is being captured.

The solution is just to disable this build-in way to just
keep the former way.

https://webrtc-review.googlesource.com/c/src/+/30902