Backport a WebRTC patch for video cropping to prevent freezes.
Fixes the problem documented in this bug:
https://bugs.chromium.org/p/webrtc/issues/detail?id=6958

where the encoder receives frame sizes it's not expecting and errors
out. The fix landed in M61:
https://codereview.webrtc.org/2936393002

This applies the critical pieces of that change to crop and scale poorly
sized frames before they hit the encoder.

Verified the fix with Slack video calling. It's easy to reproduce by
requesting a 641x480 getUserMedia stream from the camera. The logs will
be filled with this error:
```
Incoming frame doesn't match set resolution. Dropping.
```