Run `gn` via depot_tools to ensure stable .ninja file output
The `toolchain.ninja` file contains commands that invoke Python. It is
important to ensure that GN always detects the same path to the Python
executable, otherwise when Ninja calls it to re-generate .ninja files, it
creates different toolchain command lines which results in rebuild of
everything. This side effect can cause double duration of the static_library
build.