Store patches meta info in patches configs
Use a patches config to defines order of patches,
specify a patch owner, and provide a description for a patch.

It should help us to avoid order collisions, make merge conflicts explicit,
and help find owners since git blame might not be reliable
because of frequent changes during Chromium upgrades.