Add patch to fix linking error on arm64
Currently we are seeing this linking error for arm64:
./libskia.so: undefined reference to `FT_Get_Var_Design_Coordinates'
./libskia.so: undefined reference to `FT_Set_Default_Properties'

According to https://codereview.chromium.org/2892163003, removing
SK_FREETYPE_MINIMUM_RUNTIME_VERSION should be able to fix this, but also
according to https://crbug.com/715429, removing it would cause crash on
old Android systems.

Since we are not building for Android, let's just pick the easy fix.