Add a patch to fix a Mac screen sharing crash.
Backports the change here:
https://chromium.googlesource.com/external/webrtc/trunk/webrtc/+/07961e9d2becadcf9419cede22aa0bf3058ccaa2%5E%21/#F0

fixing the crash reported here:
https://bugs.chromium.org/p/chromium/issues/detail?id=752036
(requires a chromium.org account to view)

We've had crashes with the same stack trace reported by Slack screen
sharing users. Since we can't reproduced the crash internally, I haven't
been able to verify the fix, but by inspection it's very likely to work.

The upstream fix was released in M60 so this isn't as issue in Electron
2.x which uses M61.