chore: [gn] build icu & v8 statically in GN build
When building Electron with GN in release mode, all code is linked
statically into a single binary, including node & v8. This is in
contrast to the GYP build, in which node, v8 and icu are linked as a
shared library (due to openssl symbol conflicts).

This change adds a `is_electron_gn_build` flag which defaults to false,
allowing those libraries to be built statically in the GN build, but
be built as a shared library when libcc is to be consumed by the GYP
build of Electron.